fix: incorrect parsing of blocks produced wrong outputs
The text was not divided into blocks correctl. Particularly, texts  with
lines prefixed by the wild character * were poorly formatted because the
this character was treated as a beginning of list marker. This  behavior
was purposefuly added because in Markdown list may use the  *  character
as a bullet.

This creates dillema: how can the code know if the  user  is  trying  to
format an indented list whose bullet is *  or  a  multiline  C++  or  JS
comment block (like the following)?

    /**
     * This is a multiline comment block.
     * This is a multiline comment block.
     * This is a multiline comment block.
     *
     */

Checking the start of the comment block, which is /* in JS, would add
some complexity conditionals and extras checks in the codebase, which is
not worth it.

The solution is to drop support  for  list  items  with  bullet  *  and,
instead of that, add support for C++ and JS multiline comments.